refactor: no default, useless file
BREAKING CHANGE: Remove the `default` export from 'nostics/unplugin':

  ```diff
  -import nosticsStrip from 'nostics/unplugin/strip-transform'
  +import { nosticsStrip } from 'nostics/unplugin/strip-transform'
  ```
